+++ sklearn/mixture/__init__.py
@@ -2,8 +2,8 @@
 The :mod:`sklearn.mixture` module implements mixture modeling algorithms.
 """
 
-from .gaussian_mixture import GaussianMixture
-from .bayesian_mixture import BayesianGaussianMixture
+from ._gaussian_mixture import GaussianMixture
+from ._bayesian_mixture import BayesianGaussianMixture
 
 
 __all__ = ['GaussianMixture',
