+++ sklearn/mixture/tests/test_bayesian_mixture.py
@@ -13,8 +13,8 @@ from sklearn.utils.testing import assert_array_equal
 
 from sklearn.metrics.cluster import adjusted_rand_score
 
-from sklearn.mixture.bayesian_mixture import _log_dirichlet_norm
-from sklearn.mixture.bayesian_mixture import _log_wishart_norm
+from sklearn.mixture._bayesian_mixture import _log_dirichlet_norm
+from sklearn.mixture._bayesian_mixture import _log_wishart_norm
 
 from sklearn.mixture import BayesianGaussianMixture
